# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Classroom/Venue Management System API",
    "description": "API documentation for the Classroom/Venue Management System.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://wiman.azurewebsites.net/api/",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all users",
        "description": "Retrieve all users in the system.",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-swagger-router-controller": "Admin"
      }
    },
...
    "/admin/toggle-ban/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Toggle user's ban status",
        "description": "Ban or unban a user based on the current status.",
        "operationId": "toggleBanStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User ban status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User banned"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-swagger-router-controller": "Admin"
      }
    },
    "/admin/update-role/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Update a user's role",
        "description": "Update a user's role (e.g., admin, user, maintenance).",
        "operationId": "updateUserRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRole"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "User role updated"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "x-swagger-router-controller": "Admin"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "profileUrl": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["user", "admin", "maintenance"]
          },
          "blocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserRole": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["user", "admin", "maintenance"]
          },
          "blocked": {
            "type": "boolean"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found or invalid operation"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}